Fixes and stuff
combined /guest, /makeGuest, and /unGuest to be one command. If userID
is not specified then it toggles initiating admin, if one is entered
then it toggles that user instead. Removed unguest all option
fixed regex to allow all characters in parameter
fixed Help to display details even with wrong case
added a cooldown to several commands. This cooldown is currently 30
seconds and is on a per user basis. The command will not fire if the
user is still on cooldown.
Added a lot of facts.
Cleaned up fact list (3897 facts)