Fixes
+The 'peek' command returned too many results, e.g. "/peek 3" would
return the next 4 songs. Fixed math.
+Fixed wording on 'peek' now starts with #1 instead of #0
+'getPlaylistNextSongs' was using the queueID to iterate through the
songs array. This would result in an 'undefined' result if the current
song's queueID was higher than the next song's. I could not see a
practical reason to loop through the array so I eliminated the loop
completely and simply perform a sincle shift() to remove the currently
playing song from the array.
+'roll' was having problems using 'switch' and even though 'flip' would
have a value switch was not selecting a case and therefore was not
returning a correct response. For now i have switched it to an IF...ELSE
IF.